pyproject.toml
[project]
name = "fastapi-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.128.0",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.10.1",
    "sqladmin[full]>=0.22.0",
    "sqlmodel>=0.0.31",
]
